Remove `make install_dependencies`
We commit these into `vendor` now. `make install_dependencies` was from
the time when you had to run `go get ...`, which we no longer need.

Further, we definitely shouldn't run `dep` from `make compile` since we
want to compile the repo *exactly as it is now*, *not* mess around with
dependencies.

For example, imagine we've just unpacked a tarball of a
specific release, vendor directory and all. We know exactly what code is
in that tarball (and checked the sig). The last thing we want is for dep
to change anything.