Make warning messages clearer in `key list`
I've opted for lots of cases here to try and be as clear as possible
about which slice of strings is associated with each `Key Warning`.

I thought about refactoring out the 'key type', based on whether the
warning had a `SubkeyId` or not, but think that just obscured what
was going on.