add openpgpdefs subpackage
This defines constants from RFC 4880 and RFC 5881 for symmetric key
algorithms, hash algorithms and compression algorithms.

The reason for defining these is that openpgp.packet is incomplete: for
example it doesn't define any bytes that it can't support (like BZIP2).

That matters because we want to be able to *test* for BZIP2 in imported
keys to warn that it's not usable.