Backup all modified keys on `key rotate`
This ensures the user will always have a backup to disk of their latest
keys. Note: if a user imports a key from GPG into Fluidkeys and
Fluidkeys subsequently makes any changes to this key, it'll then go on
to back this up to disk, too. It does this automatically.