re-write PgpKey.Generate based on openpgp.NewEntity
the reason for abandoning openpgp.NewEntity is that it is unable to make
a primary key with a different key size than the public key.

further, we have a lot of the functionality ourselves already
(CreateNewEncryptionSubkey) so it's more consistent to just use that.

note that we never generate small 1024-bit keys for testing any more,
instead we use the mockRandom method. this allows us to really test that
the code is working with the correct key sizes.