remove golang.org/x/sys from Gopkg.toml
This was the warning I was getting:

```
Warning: the following project(s) have [[constraint]] stanzas in
Gopkg.toml:

  ✗  golang.org/x/sys

However, these projects are not direct dependencies of the current
project: they are not imported in any .go files, nor are they in the
'required' list in Gopkg.toml. Dep only applies [[constraint]] rules to
direct dependencies, so these rules will have no effect.
```