add ISSUE_TEMPLATE.md, signposting our Trello
I've made this as an issue template so that users would-be bug reporters
see the link to Trello inline, right after they start to create a new
issue.

I prefer this to asking reporters to go and read 2 separate documents
(contributing guide), since we don't have anything more to say yet.