Don't support compile from clone outside GOPATH
I tried for a long time to ensure `make compile` works *even if* you've
cloned the repo outside your GOPATH. This has caused a lot of problems
and I'm giving up.

This was to make it easy for external compile scripts like Jenkins and
Homebrew. I suspect this is going to break our deployment, so we'll have
to figure out how to fix them separately.