fix for loop pointer bug
When you're ranging over a slice if pointers like this:

`for i, value := range keys`

it's a bug to take the address of `&value` (and store it for later)

The reason is that `value` is a variable that's defined once: it has the
same address (`&value`) for every iteration through the for loop.

The *content* of `value` changes, the *address* doesn't.

So any pointer you make (by taking the address) will always point at the
same content (the one in the *last* iteration of the loop).

Instead, you can take a pointer back into the original slice based on
its index.