introduce containsDisallowedRune
containsDisallowedRune takes a given input string and returns true if
it contains a rune that we're choosing to block, namely all terminal
control codes, *except* line feed and carriage return. We're allowing
these two codes, since if I want to send a secret to another person,
I'd expect the content to retain it's original format with line breaks.