team create: search for existing keys for team members
At the moment, we just visually represent whether we find a
matching key for a given email address with a checkbox response.

We're also choosing to skip members for whom gpg finds multiple
keys. In the future we may want to add them to the roster, but
commented out.

Next I'll use this to write out `roster.toml`.