factor requests back into parent function
I'll later want to loop over these requests, one-by-one, to delete
them (if-and-only-if we successfully verify and upload the new
roster as part of `teamAuthorize`).

In doing this I also fixed a bug which was we were previously
returning an `exitCode` of 1 if there were no requests to join the
team. I think we really want to return 0 in that instance.