export table.FormatKeyTablePrimaryInstruction
Previously we always printed the primary instruction straight
after the table of keys. We're not going to do this on `status`
so have access to the function outside of the package will be
necessary.

Pulling `makePrimaryInstruction` out of `FormatKeyTable` resulted
in making it wrap `makeTable` (and therefore make `makeTable`
redundant).