improve tests & error if unhandled item/blank verb
I've tried to make these tests less confusing. In doing so, I
realised it'd be good if `RecordLast` threw an error if it's
passed an item it doesn't know how to handle, or an empty verb
string. Tests FTW!

The deleted tests at the end of this commit felt redundent now,
as the other tests in the suite write to an empty database, or
a database with existing values in by necessity!