simplify mocking interfaces: PgpKey & GnuPG
previously, we were using several different interfaces to mock out
different bits of the PgpKey or the GnuPG wrapper:

* ExportPrivateKeyInterface
* ImportArmoredKeyInterface

each interface had its own associated mock type for the tests.

then I wanted to also mock TrustUltimately, which would require adding
*another* interface and *another* mock type.

Instead, I've simplified it so we've got a `PgpKeyInterface` and a
`GnuPGInterface` which cover as many functions as we're interested in
mocking.

(Note that there's 1 more interface called `LoaderInterface` which wraps
the pgpkey.Loader struct, which *in turn* wraps the package function
`pgpkey.LoadFrom[etc]`