don't shorten long expiries
(but do still add expiry if it's missing)

since we're not aggressively rotating keys any more, it doesn't make
sense to shorten a long expiry.

plus, if someone imports a key from gpg and they've got 2 years on it, it's
not a great first impression that fluidkeys immediately wants to mess
with its (perfectly reasonably) expiry.