on launchd enable, notify before editing crontab
if we've just enabled launchd for the first time, we try and remove
legacy crontab lines.

If and only if the cron lines were actually present, print a permissions
primer and checkbox before trying to remove them.

This prevents the situation where someone types `fk` and all they see is
a permissions dialog asking to "administer the computer". At least
they'll get some heads-up about why.